use a long instead of an int in case of overflow issues
git-svn-id: https://ut.svn.beanstalkapp.com/ngfw/work/src@41741 8ebf8b4e-06f3-0310-961f-89f74618be54

+++ openvpn/src/com/untangle/node/openvpn/OpenVpnNodeImpl.java
@@ -597,9 +597,9 @@ public class OpenVpnNodeImpl extends NodeBase implements OpenVpnNode
                             continue;
                         }
                     
-                        int i;
+                        long i;
                         try {
-                            i = Integer.parseInt( parts[1] );
+                            i = Long.parseLong( parts[1] );
                         } catch ( Exception e) {
                             logger.warn("Malformed int in openvpn status: " + currentLine );
                             continue;
@@ -622,9 +622,9 @@ public class OpenVpnNodeImpl extends NodeBase implements OpenVpnNode
                             continue;
                         }
                     
-                        int i;
+                        long i;
                         try {
-                            i = Integer.parseInt( parts[1] );
+                            i = Long.parseLong( parts[1] );
                         } catch ( Exception e) {
                             logger.warn("Malformed int in openvpn status: " + currentLine );
                             continue;
